requirements.txt
chardet>=2.3.0
GitPython>=1.0.2
python-Levenshtein>=0.12.0
PyYAML>=3.11
tqdm>=3.4.0
whatthepatch>=0.0.4
